moved convolution task stuff to model module
