+++ benchmark/build.gradle
@@ -31,6 +31,9 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
+    aaptOptions {
+        noCompress ''
+    }
 }
 
 dependencies {
