+++ model/build.gradle
@@ -22,7 +22,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
-
+    aaptOptions {
+        noCompress ''
+    }
 }
 
 dependencies {
